col selector refactor
