+++ query/src/main/scala/filodb/query/exec/PeriodicSamplesMapper.scala
@@ -15,33 +15,6 @@ import filodb.query.RangeFunctionId._
 import filodb.query.exec.rangefn._
 import filodb.query.util.IndexedArrayQueue
 
-
-final case object PeriodicSamplesMapper {
-  def downsampleColsFromRangeFunction(f: Option[RangeFunctionId]): Seq[String] = {
-    f match {
-      case None                   => Seq("sum", "count") // to calculate last average
-      case Some(Rate)             => Seq("value")
-      case Some(Irate)            => Seq("value")
-      case Some(Increase)         => Seq("value")
-      case Some(Resets)           => Seq("value")
-      case Some(CountOverTime)    => Seq("count")
-      case Some(Changes)          => Seq("avg")
-      case Some(Delta)            => Seq("avg")
-      case Some(Idelta)           => Seq("avg")
-      case Some(Deriv)            => Seq("avg")
-      case Some(HoltWinters)      => Seq("avg")
-      case Some(PredictLinear)    => Seq("avg")
-      case Some(SumOverTime)      => Seq("sum")
-      case Some(AvgOverTime)      => Seq("sum", "count")
-      case Some(StdDevOverTime)   => Seq("avg")
-      case Some(StdVarOverTime)   => Seq("avg")
-      case Some(QuantileOverTime) => Seq("avg")
-      case Some(MinOverTime)      => Seq("min")
-      case Some(MaxOverTime)      => Seq("max")
-    }
-  }
-}
-
 /**
   * Transforms raw reported samples to samples with
   * regular interval from start to end, with step as
